moved CommandPlugin
