+++ .gitignore
@@ -9,7 +9,8 @@ logs.txt
 log.txt
 
 settings_prod.py
-*private
+plugins/custom/*
+!plugins/custom/__init__.py
 
 *.json
 *.notjson
