+++ handler/template_plugin_command.py
@@ -1,4 +1,4 @@
-from handler.base_plugin_command import CommandPlugin
+from handler.base_plugin import CommandPlugin
 
 
 class xPlugin(CommandPlugin):
