+++ plugins/content/content_birthday.py
@@ -1,6 +1,6 @@
 import datetime
 
-from handler.base_plugin_command import CommandPlugin
+from handler.base_plugin import CommandPlugin
 from utils import plural_form, age
