+++ plugins/content/content_calculation/calculator_plugin.py
@@ -1,4 +1,4 @@
-from handler.base_plugin_command import CommandPlugin
+from handler.base_plugin import CommandPlugin
 from .calculator import Calculator
 
 class CalculatorPlugin(CommandPlugin):
