+++ plugins/control/control_dispatch.py
@@ -1,4 +1,4 @@
-from handler.base_plugin_command import CommandPlugin
+from handler.base_plugin import CommandPlugin
 from skevk import Wait, upload_photo
 
 import asyncio
