+++ plugins/miscellaneous/misc_counter.py
@@ -1,4 +1,4 @@
-from handler.base_plugin_command import CommandPlugin
+from handler.base_plugin import CommandPlugin
 
 
 class CounterPlugin(CommandPlugin):
