+++ plugins/multiuser/chat_members.py
@@ -1,6 +1,6 @@
 import random
 
-from handler.base_plugin_command import CommandPlugin
+from handler.base_plugin import CommandPlugin
 
 
 class MembersPlugin(CommandPlugin):
