+++ plugins/outsource/outsource_weather.py
@@ -1,4 +1,4 @@
-from handler.base_plugin_command import CommandPlugin
+from handler.base_plugin import CommandPlugin
 
 from geopy.geocoders import Photon, Yandex, Nominatim
 import aiohttp
